dom/workers/test/serviceworkers/test_get_serviced_enumerate.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 982726 - test get_serviced </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">
  var registration;
  var opened = [];
  var results = [];

  function start() {
    dump("~*~*~*~*~*~*~*~*~*~*~ About to register\n");
    return navigator.serviceWorker.register("get_serviced_worker_enumerate.js",
                                            { scope: "./sw_clients/" })
      .then((swr) => registration = swr);
  }

  function closeAndUnregister() {
    var content = document.getElementById("content");
    ok(content, "Parent exists.");

    while (opened.length) {
      content.removeChild(opened.pop());
    }

    return registration.unregister().then(function(result) {
      ok(result, "Unregister should return true.");
    });
  }

  function openClient() {
    var p = new Promise(function(res, rej) {

      var content = document.getElementById("content");
      ok(content, "Parent exists.");

      var iframe = document.createElement("iframe");
      iframe.setAttribute('src', "sw_clients/service_worker_controlled.html");

      content.appendChild(iframe);

      window.onmessage = function(e) {
        if (e.data === "READY") {
          opened.push(iframe);
          res();
        }
      }
    });
    return p;
  }

  function testGetServiced() {
    var count = 0;
    return new Promise(function (res, rej) {
      if (!registration.active) {
        ok(false, "Did not have active worker");
        res();
        return;
      }
      registration.active.postMessage("getServiced");
      // wait for response
      window.onmessage = function(msg) {
        results[msg.data] = "ok";
        count = count + 1;
        if (count == opened.length) {
          for (i = 0; i < count; i++) {
            ok(results[i] == "ok", "Client received the message.");
          }
          res();
        }
      }
    });
  }

  function runTest() {
    start()
      .then(openClient).then(openClient)
      .then(openClient).then(openClient)
      .then(testGetServiced)
      .then(closeAndUnregister)
      .catch(function(e) {
        ok(false, "Some test failed with error " + e);
      }).then(function() {
        SimpleTest.finish();
      });
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
  ]}, runTest);
</script>
</pre>
</body>
</html>
